exoplayer: upd sabr classes; media notification fix

+++ exoplayer-amzn-2.10.6/library/sabr/src/main/java/com/google/android/exoplayer2/source/sabr/parser/models/SelectedFormat.java
@@ -14,7 +14,7 @@ public class SelectedFormat {
     public final FormatSelector formatSelector;
     public int totalSegments;
     public final boolean discard;
-    public int sequenceLmt = -1;
+    public long sequenceLmt = -1;
     public Segment currentSegment;
     public Segment initSegment;
     public final List<ConsumedRange> consumedRanges = new ArrayList<>();
